refactoring UNetStructure.py, UNet.py
