+++ script/model/sklearn_like_model/UNet.py
@@ -2,10 +2,9 @@ from tqdm import trange
 import numpy as np
 from script.data_handler.Base.BaseDataset import BaseDataset
 from script.model.sklearn_like_model.BaseModel import BaseModel
-from script.model.sklearn_like_model.Base_net_structure import Base_net_structure
 from script.model.sklearn_like_model.Mixin import Xs_MixIn, Ys_MixIn, supervised_trainMethodMixIn, predictMethodMixIn, \
     predict_probaMethodMixIn, scoreMethodMixIn, supervised_metricMethodMixIn
-from script.util.Stacker import Stacker
+from script.model.sklearn_like_model.net_structure.UNetStructure import UNetStructure
 from script.util.tensor_ops import *
 
 
@@ -66,52 +65,6 @@ def pixel_wise_softmax(output_map, name='pixel_wise_softmax', reuse=False):
         return exponential_map / normalize
 
 
-class Unet_structure(Base_net_structure):
-
-    def __init__(self, Xs, level=4, n_classes=2, reuse=False, name='Unet'):
-        super().__init__(reuse, name)
-
-        self.X = Xs
-        self.level = level
-        self.n_classes = n_classes
-        self.stacker = Stacker(Xs)
-
-    def build(self):
-        self.logit, self.proba = self.Unet_recursion_build(
-            self.stacker, self.level, self.n_classes, self.reuse, self.name)
-
-    @staticmethod
-    def Unet_recursion_build(stacker, level=4, n_classes=2, reuse=False, name='Unet'):
-        def _Unet_recursion(stacker, n_channel, level):
-            if level == 0:
-                stacker.conv_block(n_channel, CONV_FILTER_3311, relu)
-                stacker.conv_block(n_channel, CONV_FILTER_3311, relu)
-            else:
-                # encode
-                stacker.conv_block(n_channel, CONV_FILTER_3311, relu)
-                stacker.conv_block(n_channel, CONV_FILTER_3311, relu)
-                concat = stacker.last_layer
-                stacker.max_pooling(CONV_FILTER_2211)
-
-                stacker = _Unet_recursion(stacker, n_channel * 2, level - 1)
-
-                # decode
-                stacker.upscale_2x_block(n_channel, CONV_FILTER_2211, relu)
-                stacker.concat(concat, axis=3)
-                stacker.conv_block(n_channel, CONV_FILTER_3311, relu)
-                stacker.conv_block(n_channel, CONV_FILTER_3311, relu)
-
-            return stacker
-
-        with tf.variable_scope(name, reuse=reuse):
-            stacker = _Unet_recursion(stacker, n_channel=64, level=level)
-            stacker.conv_block(n_classes, CONV_FILTER_3311, relu)
-            logit = stacker.last_layer
-            proba = pixel_wise_softmax(logit)
-
-        return logit, proba
-
-
 class UNet(
     BaseModel,
     Xs_MixIn,
@@ -156,7 +109,7 @@ class UNet(
         self.Xs = tf.placeholder(tf.float32, self.Xs_shape, name='Xs')
         self.Ys = tf.placeholder(tf.float32, self.Ys_shape, name='Ys')
 
-        self.Unet_structure = Unet_structure(self.Xs, level=self.stage)
+        self.Unet_structure = UNetStructure(self.Xs, level=self.stage)
         self.Unet_structure.build()
         self._logit = self.Unet_structure.logit
         self._proba = self.Unet_structure.proba
